ai-customer-support-bot/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
httpx>=0.24.0
jinja2>=3.0.0
python-multipart>=0.0.5
